+++ src/SRISYS.Web/wwwroot/js/controllers/system/view-suppliers-controller.js
@@ -1,5 +1,5 @@
 ﻿(function (module) {
-    var viewSuppliersController = function ($q, supplierService, loadingService) {
+    var viewSuppliersController = function ($q, supplierService, utils) {
         var vm = this;
         vm.focus = true;
         vm.currentPage = 1;
@@ -25,11 +25,11 @@
 
         // Methods
         vm.fetchSuppliers = function () {
-            loadingService.showLoading();
+            utils.showLoading();
 
             supplierService.searchSuppliers(vm.filters)
-                .then(processSupplierList, onFetchError)
-                .finally(hideLoading);
+                .then(processSupplierList, utils.onError)
+                .finally(utils.hideLoading);
         };
 
         vm.clearFilter = function () {
@@ -38,27 +38,25 @@
             vm.focus = true;
 		};
 
-		vm.delete = function (id) {
+        vm.delete = function (id) {
+            if (!utils.showConfirmMessage("Are you sure you want to delete this supplier?")) { return; }
+
 			supplierService.deleteSupplier(id)
-				.then((response) => { vm.fetchSuppliers(); }, onFetchError)
+				.then((response) => { vm.fetchSuppliers(); }, utils.onError)
 				.finally(utils.hideLoading);
 		};
+
+        // Paging
+        vm.changePage = function () {
+            vm.fetchSuppliers();
+        };
 
         var processSupplierList = function (response) {
             angular.copy(response.data, vm.summaryResult);
         };
 
-        var onFetchError = function (error) {
-            toastr.error("There was an error processing your requests.", "error");
-            console.log(error);
-        };
-
-        var hideLoading = function () {
-            loadingService.hideLoading();
-        };
-
         var loadAll = function () {
-            loadingService.showLoading();
+            utils.showLoading();
 
             var requests = {
                 supplier: supplierService.searchSuppliers(vm.filters)
@@ -67,8 +65,8 @@
             $q.all(requests)
                 .then((responses) => {
                     processSupplierList(responses.supplier);
-                }, onFetchError)
-                .finally(hideLoading);
+                }, utils.onError)
+                .finally(utils.hideLoading);
         };
 
         $(function () {
@@ -78,6 +76,6 @@
         return vm;
     };
 
-    module.controller("viewSuppliersController", ["$q", "supplierService", "loadingService", viewSuppliersController]);
+    module.controller("viewSuppliersController", ["$q", "supplierService", "utils", viewSuppliersController]);
 
 })(angular.module("srisys-app"));
